chore(release): moved the git tags logic into it's own file
- move the instance into release-graph class
- pass resolveGitTags as params
