+++ packages/nx/src/command-line/release/changelog.ts
@@ -1,15 +1,11 @@
 import * as chalk from 'chalk';
 import { prompt } from 'enquirer';
 import { readFileSync, rmSync, writeFileSync } from 'node:fs';
-import { ReleaseType, prerelease } from 'semver';
+import { prerelease } from 'semver';
 import { dirSync } from 'tmp';
 import type { DependencyBump } from '../../../release/changelog-renderer';
 import { NxReleaseConfiguration, readNxJson } from '../../config/nx-json';
-import {
-  FileData,
-  ProjectFileMap,
-  ProjectGraphProjectNode,
-} from '../../config/project-graph';
+import { ProjectGraphProjectNode } from '../../config/project-graph';
 import { FsTree, Tree } from '../../generators/tree';
 import {
   createFileMapUsingProjectGraph,
@@ -22,6 +18,22 @@ import { isCI } from '../../utils/is-ci';
 import { output } from '../../utils/output';
 import { joinPathFragments } from '../../utils/path';
 import { workspaceRoot } from '../../utils/workspace-root';
+import {
+  createChangesFromCommits,
+  createFileToProjectMap,
+  filterHiddenChanges,
+  mapCommitToChange,
+} from './changelog/commit-utils';
+import {
+  filterVersionPlansByCommitRange,
+  resolveChangelogFromSHA,
+  resolveWorkspaceChangelogFromSHA,
+} from './changelog/version-plan-filtering';
+import {
+  ChangelogChange,
+  createChangesFromGroupVersionPlans,
+  createChangesFromProjectsVersionPlans,
+} from './changelog/version-plan-utils';
 import { ChangelogOptions } from './command-object';
 import {
   NxReleaseConfig,
@@ -39,11 +51,8 @@ import {
 } from './config/version-plans';
 import {
   GitCommit,
-  Reference,
   getCommitHash,
-  getFirstGitCommit,
   getGitDiff,
-  getLatestGitTagForPattern,
   gitAdd,
   gitPush,
   gitTag,
@@ -55,6 +64,7 @@ import { printAndFlushChanges } from './utils/print-changes';
 import { printConfigAndExit } from './utils/print-config';
 import { ReleaseGraph, createReleaseGraph } from './utils/release-graph';
 import { createRemoteReleaseClient } from './utils/remote-release-clients/remote-release-client';
+import type { CheckAllBranchesWhen } from './utils/repository-git-tags';
 import { resolveChangelogRenderer } from './utils/resolve-changelog-renderer';
 import { resolveNxJsonConfigErrorMessage } from './utils/resolve-nx-json-error-message';
 import {
@@ -72,22 +82,6 @@ import {
   areAllVersionPlanProjectsFiltered,
   validateResolvedVersionPlansAgainstFilter,
 } from './utils/version-plan-utils';
-import {
-  ChangelogChange,
-  createChangesFromGroupVersionPlans,
-  createChangesFromProjectsVersionPlans,
-} from './changelog/version-plan-utils';
-import {
-  createChangesFromCommits,
-  createFileToProjectMap,
-  filterHiddenChanges,
-  mapCommitToChange,
-} from './changelog/commit-utils';
-import {
-  filterVersionPlansByCommitRange,
-  resolveWorkspaceChangelogFromSHA,
-  resolveChangelogFromSHA,
-} from './changelog/version-plan-filtering';
 
 export interface NxReleaseChangelogResult {
   workspaceChangelog?: {
@@ -244,6 +238,7 @@ export function createAPI(
       args,
       nxReleaseConfig,
       useAutomaticFromRef,
+      resolveRepositoryTags: releaseGraph.resolveRepositoryTags,
     });
 
     // Filter version plans based on resolveVersionPlans option
@@ -372,7 +367,7 @@ export function createAPI(
       pattern: string,
       templateValues: Record<string, string>,
       preid: string | undefined,
-      checkAllBranchesWhen: boolean | string[],
+      checkAllBranchesWhen: CheckAllBranchesWhen,
       requireSemver: boolean,
       strictPreid: boolean
     ): Promise<string | null> => {
@@ -383,6 +378,7 @@ export function createAPI(
         fromRef: args.from,
         tagPattern: pattern,
         tagPatternValues: templateValues,
+        resolveRepositoryTags: releaseGraph.resolveRepositoryTags,
         checkAllBranchesWhen,
         preid,
         requireSemver,
