+++ packages/nx/src/command-line/release/utils/git.spec.ts
@@ -1,9 +1,9 @@
 import {
   extractReferencesFromCommit,
   getLatestGitTagForPattern,
-  RepoGitTags,
   sanitizeProjectNameForGitTag,
 } from './git';
+import { RepoGitTags } from './repository-git-tags';
 
 jest.mock('./exec-command', () => ({
   execCommand: jest.fn(() =>
@@ -151,9 +151,12 @@ See merge request nx-release-test/nx-release-test!2`,
   });
 
   describe('getLatestGitTagForPattern', () => {
+    // Create a mock resolveTags function that uses the mocked execCommand
+    const mockRepoGitTags = RepoGitTags.create();
+    const mockResolveTags = mockRepoGitTags.resolveTags;
+
     afterEach(() => {
       jest.clearAllMocks();
-      RepoGitTags.instance.clean();
     });
 
     describe('when releaseTagPatternStrictPreid is false', () => {
@@ -352,6 +355,7 @@ See merge request nx-release-test/nx-release-test!2`,
               projectName,
               releaseGroupName,
             },
+            mockResolveTags,
             {
               requireSemver,
               preid: preid,
@@ -451,6 +455,7 @@ See merge request nx-release-test/nx-release-test!2`,
             {
               projectName,
             },
+            mockResolveTags,
             {
               preid: preid,
               requireSemver: true,
@@ -476,6 +481,7 @@ See merge request nx-release-test/nx-release-test!2`,
         {
           projectName: 'my-lib-1',
         },
+        mockResolveTags,
         {
           requireSemver: true,
           strictPreid: false,
@@ -490,6 +496,7 @@ See merge request nx-release-test/nx-release-test!2`,
         {
           projectName: 'my-lib-1',
         },
+        mockResolveTags,
         {
           requireSemver: true,
           strictPreid: false,
